DeviceId wrong convertion to Uint16  in the schedule editor
git-svn-id: https://svn.code.sf.net/p/yetanotherbacnetexplorer/code/trunk@278 8324893d-a7a8-4806-89d4-2f8fea5c7913

+++ Yabe/ScheduleDisplay.cs
@@ -689,7 +689,7 @@ namespace Yabe
                 uint ArrayIdx = ASN1.BACNET_ARRAY_ALL;
 
                 if (Reference_Device.Text != "")
-                    device = new BacnetObjectId(BacnetObjectTypes.OBJECT_DEVICE, Convert.ToUInt16(Reference_Device.Text));
+                    device = new BacnetObjectId(BacnetObjectTypes.OBJECT_DEVICE, Convert.ToUInt32(Reference_Device.Text));
                 if (Reference_Array.Text != "")
                     ArrayIdx = Convert.ToUInt16(Reference_Array.Text);
